using a modal to do updates now. allows for a commit message to be passed in
